+++ SClass-Common/src/main/kotlin/com/sclass/common/config/CorsConfig.kt
@@ -1,0 +1,27 @@
+package com.sclass.common.config
+
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.CorsRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration
+class CorsConfig(
+    @param:Value("\${cors.allow.origins:}") private val allowedOrigins: String,
+) : WebMvcConfigurer {
+    override fun addCorsMappings(registry: CorsRegistry) {
+        val origins =
+            allowedOrigins
+                .split(",")
+                .map(String::trim)
+                .filter(String::isNotEmpty)
+        if (origins.isEmpty()) return
+
+        registry
+            .addMapping("/**")
+            .allowedOrigins(*origins.toTypedArray())
+            .allowedMethods("GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS")
+            .allowedHeaders("*")
+            .allowCredentials(true)
+    }
+}
